+++ heat/core/manipulations.py
@@ -11,10 +11,11 @@ from . import linalg
 from . import stride_tricks
 from . import tiling
 from . import types
-from . import operations
+from . import _operations
 
 
 __all__ = [
+    "column_stack",
     "concatenate",
     "diag",
     "diagonal",
@@ -27,24 +28,117 @@ __all__ = [
     "reshape",
     "resplit",
     "rot90",
+    "row_stack",
     "shape",
     "sort",
     "squeeze",
+    "stack",
     "topk",
     "unique",
     "vstack",
 ]
 
 
-def concatenate(arrays, axis=0):
+def column_stack(arrays):
     """
-    Join 2 arrays along an existing axis.
+    Stack 1-D or 2-D ``DNDarray``s as columns into a 2-D ``DNDarray``.
+    If the input arrays are 1-D, they will be stacked as columns. If they are 2-D,
+    they will be concatenated along the second axis.
 
     Parameters
     ----------
-    arrays: tuple of 2 DNDarrays
+    arrays : Sequence[DNDarrays,...]
+
+    Raises
+    ------
+    ValueError
+        If arrays have more than 2 dimensions
+
+    Returns
+    -------
+    DNDarray
+
+    Note
+    ----
+    All ``DNDarray``s in the sequence must have the same number of rows.
+    All ``DNDarray``s must be split along the same axis! Note that distributed
+    1-D arrays (``split = 0``) by default will be transposed into distributed
+    column arrays with ``split == 1``.
+
+    Examples
+    --------
+    >>> # 1-D tensors
+    >>> a = ht.array([1, 2, 3])
+    >>> b = ht.array([2, 3, 4])
+    >>> ht.column_stack((a, b))._DNDarray__array
+    tensor([[1, 2],
+        [2, 3],
+        [3, 4]])
+    >>> # 1-D and 2-D tensors
+    >>> a = ht.array([1, 2, 3])
+    >>> b = ht.array([[2, 5], [3, 6], [4, 7]])
+    >>> c = ht.array([[7, 10], [8, 11], [9, 12]])
+    >>> ht.column_stack((a, b, c))._DNDarray__array
+    tensor([[ 1,  2,  5,  7, 10],
+            [ 2,  3,  6,  8, 11],
+            [ 3,  4,  7,  9, 12]])
+    >>> # distributed DNDarrays, 3 processes
+    >>> a = ht.arange(10, split=0).reshape((5, 2))
+    >>> b = ht.arange(5, 20, split=0).reshape((5, 3))
+    >>> c = ht.arange(20, 40, split=0).reshape((5, 4))
+    >>> ht_column_stack((a, b, c))._DNDarray__array
+    [0/2] tensor([[ 0,  1,  5,  6,  7, 20, 21, 22, 23],
+    [0/2]         [ 2,  3,  8,  9, 10, 24, 25, 26, 27]], dtype=torch.int32)
+    [1/2] tensor([[ 4,  5, 11, 12, 13, 28, 29, 30, 31],
+    [1/2]         [ 6,  7, 14, 15, 16, 32, 33, 34, 35]], dtype=torch.int32)
+    [2/2] tensor([[ 8,  9, 17, 18, 19, 36, 37, 38, 39]], dtype=torch.int32)
+    >>> # distributed 1-D and 2-D DNDarrays, 3 processes
+    >>> a = ht.arange(5, split=0)
+    >>> b = ht.arange(5, 20, split=1).reshape((5, 3))
+    >>> ht_column_stack((a, b))._DNDarray__array
+    [0/2] tensor([[ 0,  5],
+    [0/2]         [ 1,  8],
+    [0/2]         [ 2, 11],
+    [0/2]         [ 3, 14],
+    [0/2]         [ 4, 17]], dtype=torch.int32)
+    [1/2] tensor([[ 6],
+    [1/2]         [ 9],
+    [1/2]         [12],
+    [1/2]         [15],
+    [1/2]         [18]], dtype=torch.int32)
+    [2/2] tensor([[ 7],
+    [2/2]         [10],
+    [2/2]         [13],
+    [2/2]         [16],
+    [2/2]         [19]], dtype=torch.int32)
+    """
+    arr_dims = list(array.ndim for array in arrays)
+    # sanitation, arrays can be 1-d or 2-d, see sanitation module #468
+    over_dims = [i for i, j in enumerate(arr_dims) if j > 2]
+    if len(over_dims) > 0:
+        raise ValueError("Arrays must be 1-D or 2-D")
+    if arr_dims.count(1) == len(arr_dims):
+        # all arrays are 1-D, stack
+        return stack(arrays, axis=1)
+    else:
+        if arr_dims.count(1) > 0:
+            arr_1d = [i for i, j in enumerate(arr_dims) if j == 1]
+            # 1-D arrays must be columns
+            arrays = list(arrays)
+            for ind in arr_1d:
+                arrays[ind] = arrays[ind].reshape((1, arrays[ind].size)).T
+        return concatenate(arrays, axis=1)
+
+
+def concatenate(arrays, axis=0):
+    """
+    Join a sequence of arrays along an existing axis.
+
+    Parameters
+    ----------
+    arrays : Sequence[DNDarrays,...]
         The arrays must have the same shape, except in the dimension corresponding to axis (the first, by default).
-    axis: int, optional
+    axis : int, optional
         The axis along which the arrays will be joined. Default is 0.
 
     Returns
@@ -788,20 +882,20 @@ def hstack(tup):
     --------
     >>> a = ht.array((1,2,3))
     >>> b = ht.array((2,3,4))
-    >>> ht.hstack((a,b))
-    [0] tensor([1, 2, 3, 2, 3, 4])
-    [1] tensor([1, 2, 3, 2, 3, 4])
+    >>> ht.hstack((a,b))._DNDarray__array
+    [0/1] tensor([1, 2, 3, 2, 3, 4])
+    [1/1] tensor([1, 2, 3, 2, 3, 4])
     >>> a = ht.array((1,2,3), split=0)
     >>> b = ht.array((2,3,4), split=0)
-    >>> ht.hstack((a,b))
-    [0] tensor([1, 2, 3])
-    [1] tensor([2, 3, 4])
+    >>> ht.hstack((a,b))._DNDarray__array
+    [0/1] tensor([1, 2, 3])
+    [1/1] tensor([2, 3, 4])
     >>> a = ht.array([[1],[2],[3]], split=0)
     >>> b = ht.array([[2],[3],[4]], split=0)
-    >>> ht.hstack((a,b))
-    [0] tensor([[1, 2],
-    [0]         [2, 3]])
-    [1] tensor([[3, 4]])
+    >>> ht.hstack((a,b))._DNDarray__array
+    [0/1] tensor([[1, 2],
+    [0/1]         [2, 3]])
+    [1/1] tensor([[3, 4]])
     """
     tup = list(tup)
     axis = 1
@@ -1431,6 +1525,193 @@ def squeeze(x, axis=None):
     )
 
 
+def stack(arrays, axis=0, out=None):
+    """
+    Join a sequence of ``DNDarray``s along a new axis.
+
+    The ``axis`` parameter specifies the index of the new axis in the dimensions of the result.
+    For example, if ``axis=0``, the arrays will be stacked along the first dimension; if ``axis=-1``,
+    they will be stacked along the last dimension. See Notes below for split semantics.
+
+    Parameters
+    ----------
+    arrays : Sequence[DNDarrays,...]
+        Each DNDarray must have the same shape, must be split along the same axis, and must be balanced.
+    axis : int, optional
+        The axis in the result array along which the input arrays are stacked.
+    out : DNDarray, optional
+        If provided, the destination to place the result. The shape and split axis must be correct, matching
+        that of what stack would have returned if no out argument were specified (see Notes below).
+
+    Raises
+    ------
+    TypeError
+        If arrays in sequence are not ``DNDarray``s, or if their ``dtype`` attribute does not match.
+    ValueError
+        If ``arrays`` contains less than 2 ``DNDarray``s.
+    ValueError
+        If the ``DNDarray``s are of different shapes, or if they are split along different axes (``split`` attribute).
+    RuntimeError
+        If the ``DNDarrays`` reside of different devices, or if they are unevenly distributed across ranks (method ``is_balanced()`` returns ``False``)
+
+    Returns
+    -------
+    DNDarray
+
+    Notes
+    -----
+    Split semantics: :func:`stack` requires that all arrays in the sequence be split along the same dimension.
+    After stacking, the data are still distributed along the original dimension, however a new dimension has been added at `axis`,
+    therefore:
+
+    - if :math:`axis <= split`, output will be distributed along :math:`split+1`
+
+    - if :math:`axis > split`, output will be distributed along `split`
+
+    Examples
+    --------
+    >>> a = ht.arange(20).reshape(4, 5)
+    >>> b = ht.arange(20, 40).reshape(4, 5)
+    >>> ht.stack((a,b), axis=0)._DNDarray__array
+    tensor([[[ 0,  1,  2,  3,  4],
+             [ 5,  6,  7,  8,  9],
+             [10, 11, 12, 13, 14],
+             [15, 16, 17, 18, 19]],
+
+            [[20, 21, 22, 23, 24],
+             [25, 26, 27, 28, 29],
+             [30, 31, 32, 33, 34],
+             [35, 36, 37, 38, 39]]])
+    >>> # distributed DNDarrays, 3 processes, stack along last dimension
+    >>> a = ht.arange(20, split=0).reshape(4, 5)
+    >>> b = ht.arange(20, 40, split=0).reshape(4, 5)
+    >>> ht.stack((a,b), axis=-1)._DNDarray__array
+    [0/2] tensor([[[ 0, 20],
+    [0/2]          [ 1, 21],
+    [0/2]          [ 2, 22],
+    [0/2]          [ 3, 23],
+    [0/2]          [ 4, 24]],
+    [0/2]
+    [0/2]         [[ 5, 25],
+    [0/2]          [ 6, 26],
+    [0/2]          [ 7, 27],
+    [0/2]          [ 8, 28],
+    [0/2]          [ 9, 29]]])
+    [1/2] tensor([[[10, 30],
+    [1/2]          [11, 31],
+    [1/2]          [12, 32],
+    [1/2]          [13, 33],
+    [1/2]          [14, 34]]])
+    [2/2] tensor([[[15, 35],
+    [2/2]          [16, 36],
+    [2/2]          [17, 37],
+    [2/2]          [18, 38],
+    [2/2]          [19, 39]]])
+    """
+
+    # sanitation
+    if not isinstance(arrays, (tuple, list)):
+        raise TypeError("arrays must be a list or a tuple")
+
+    if len(arrays) < 2:
+        raise ValueError("stack expects a sequence of at least 2 DNDarrays")
+
+    for i, array in enumerate(arrays):
+        if not isinstance(array, dndarray.DNDarray):
+            raise TypeError(
+                "all arrays in sequence must be DNDarrays, array {} was {}".format(i, type(array))
+            )
+
+    arrays_metadata = list(
+        [array.gshape, array.split, array.device, array.is_balanced()] for array in arrays
+    )
+    num_arrays = len(arrays)
+    # metadata must be identical for all arrays
+    if arrays_metadata.count(arrays_metadata[0]) != num_arrays:
+        shapes = list(array.gshape for array in arrays)
+        if shapes.count(shapes[0]) != num_arrays:
+            raise ValueError(
+                "All DNDarrays in sequence must have the same shape, got shapes {}".format(shapes)
+            )
+        splits = list(array.split for array in arrays)
+        if splits.count(splits[0]) != num_arrays:
+            raise ValueError(
+                "All DNDarrays in sequence must have the same split axis, got splits {}"
+                "Check out the heat.resplit() documentation.".format(splits)
+            )
+        devices = list(array.device for array in arrays)
+        if devices.count(devices[0]) != num_arrays:
+            raise RuntimeError(
+                "DNDarrays in sequence must reside on the same device, got devices {}".format(
+                    devices
+                )
+            )
+        balance = list(array.is_balanced() for array in arrays)
+        if balance.count(balance[0]) != num_arrays:
+            raise RuntimeError(
+                "DNDarrays distribution must be balanced across ranks, is_balanced() returns {}"
+                "You can balance a DNDarray with the balance_() method.".format(balance)
+            )
+    else:
+        array_shape, array_split, array_device = arrays_metadata[0][:3]
+        # extract torch tensors
+        t_arrays = list(array._DNDarray__array for array in arrays)
+        # output dtype
+        t_dtypes = list(t_array.dtype for t_array in t_arrays)
+        t_array_dtype = t_dtypes[0]
+        if t_dtypes.count(t_dtypes[0]) != num_arrays:
+            for d in range(1, len(t_dtypes)):
+                t_array_dtype = (
+                    t_array_dtype
+                    if t_array_dtype is t_dtypes[d]
+                    else torch.promote_types(t_array_dtype, t_dtypes[d])
+                )
+            t_arrays = list(t_array.type(t_array_dtype) for t_array in t_arrays)
+        array_dtype = types.canonical_heat_type(t_array_dtype)
+
+    # sanitate axis
+    axis = stride_tricks.sanitize_axis(array_shape + (num_arrays,), axis)
+
+    # output shape and split
+    stacked_shape = array_shape[:axis] + (num_arrays,) + array_shape[axis:]
+    if array_split is not None:
+        stacked_split = array_split + 1 if axis <= array_split else array_split
+    else:
+        stacked_split = None
+
+    # sanitate output
+    if out is not None:
+        if not isinstance(out, dndarray.DNDarray):
+            raise TypeError("expected out to be None or ht.DNDarray, but was {}".format(type(out)))
+        if out.dtype is not array_dtype:
+            raise TypeError("expected out to be {}, but was {}".format(array_dtype, out.dtype))
+        if out.gshape != stacked_shape:
+            raise ValueError(
+                "expected out.shape to be {}, got {}".format(out.gshape, stacked_shape)
+            )
+        if out.split is not stacked_split:
+            raise ValueError("expected out.split to be {}, got {}".format(out.split, stacked_split))
+    # end of sanitation
+
+    # stack locally
+    t_stacked = torch.stack(t_arrays, dim=axis)
+
+    # return stacked DNDarrays
+    if out is not None:
+        out._DNDarray__array = t_stacked
+        return out
+
+    stacked = dndarray.DNDarray(
+        t_stacked,
+        gshape=stacked_shape,
+        dtype=array_dtype,
+        split=stacked_split,
+        device=array_device,
+        comm=arrays[0].comm,
+    )
+    return stacked
+
+
 def unique(a, sorted=False, return_inverse=False, axis=None):
     """
     Finds and returns the unique elements of an array.
@@ -1737,6 +2018,89 @@ def resplit(arr, axis=None):
     return new_arr
 
 
+def row_stack(arrays):
+    """
+    Stack 1-D or 2-D ``DNDarray``s as rows into a 2-D ``DNDarray``.
+    If the input arrays are 1-D, they will be stacked as rows. If they are 2-D,
+    they will be concatenated along the first axis.
+
+    Parameters
+    ----------
+    arrays : Sequence[DNDarrays,...]
+
+    Raises
+    ------
+    ValueError
+        If arrays have more than 2 dimensions
+
+    Returns
+    -------
+    DNDarray
+
+    Note
+    ----
+    All ``DNDarray``s in the sequence must have the same number of columns.
+    All ``DNDarray``s must be split along the same axis!
+
+    Examples
+    --------
+    >>> # 1-D tensors
+    >>> a = ht.array([1, 2, 3])
+    >>> b = ht.array([2, 3, 4])
+    >>> ht.row_stack((a, b))._DNDarray__array
+    tensor([[1, 2, 3],
+            [2, 3, 4]])
+    >>> # 1-D and 2-D tensors
+    >>> a = ht.array([1, 2, 3])
+    >>> b = ht.array([[2, 3, 4], [5, 6, 7]])
+    >>> c = ht.array([[7, 8, 9], [10, 11, 12]])
+    >>> ht.row_stack((a, b, c))._DNDarray__array
+    tensor([[ 1,  2,  3],
+            [ 2,  3,  4],
+            [ 5,  6,  7],
+            [ 7,  8,  9],
+            [10, 11, 12]])
+    >>> # distributed DNDarrays, 3 processes
+    >>> a = ht.arange(10, split=0).reshape((2, 5))
+    >>> b = ht.arange(5, 20, split=0).reshape((3, 5))
+    >>> c = ht.arange(20, 40, split=0).reshape((4, 5))
+    >>> ht.row_stack((a, b, c))._DNDarray__array
+    [0/2] tensor([[0, 1, 2, 3, 4],
+    [0/2]         [5, 6, 7, 8, 9],
+    [0/2]         [5, 6, 7, 8, 9]], dtype=torch.int32)
+    [1/2] tensor([[10, 11, 12, 13, 14],
+    [1/2]         [15, 16, 17, 18, 19],
+    [1/2]         [20, 21, 22, 23, 24]], dtype=torch.int32)
+    [2/2] tensor([[25, 26, 27, 28, 29],
+    [2/2]         [30, 31, 32, 33, 34],
+    [2/2]         [35, 36, 37, 38, 39]], dtype=torch.int32)
+    >>> # distributed 1-D and 2-D DNDarrays, 3 processes
+    >>> a = ht.arange(5, split=0)
+    >>> b = ht.arange(5, 20, split=0).reshape((3, 5))
+    >>> ht.row_stack((a, b))._DNDarray__array
+    [0/2] tensor([[0, 1, 2, 3, 4],
+    [0/2]         [5, 6, 7, 8, 9]])
+    [1/2] tensor([[10, 11, 12, 13, 14]])
+    [2/2] tensor([[15, 16, 17, 18, 19]])
+    """
+    arr_dims = list(array.ndim for array in arrays)
+    # sanitation, arrays can be 1-d or 2-d, see sanitation module #468
+    over_dims = [i for i, j in enumerate(arr_dims) if j > 2]
+    if len(over_dims) > 0:
+        raise ValueError("Arrays must be 1-D or 2-D")
+    if arr_dims.count(1) == len(arr_dims):
+        # all arrays are 1-D, stack
+        return stack(arrays, axis=0)
+    else:
+        if arr_dims.count(1) > 0:
+            arr_1d = [i for i, j in enumerate(arr_dims) if j == 1]
+            # 1-D arrays must be row arrays
+            arrays = list(arrays)
+            for ind in arr_1d:
+                arrays[ind] = arrays[ind].reshape((1, arrays[ind].size))
+        return concatenate(arrays, axis=0)
+
+
 def vstack(tup):
     """
     Stack arrays in sequence vertically (row wise).
@@ -1761,28 +2125,27 @@ def vstack(tup):
     --------
     >>> a = ht.array([1, 2, 3])
     >>> b = ht.array([2, 3, 4])
-    >>> ht.vstack((a,b))
-    [0] tensor([[1, 2, 3],
-    [0]         [2, 3, 4]])
-    [1] tensor([[1, 2, 3],
-    [1]         [2, 3, 4]])
+    >>> ht.vstack((a,b))._DNDarray__array
+    [0/1] tensor([[1, 2, 3],
+    [0/1]         [2, 3, 4]])
+    [1/1] tensor([[1, 2, 3],
+    [1/1]         [2, 3, 4]])
     >>> a = ht.array([1, 2, 3], split=0)
     >>> b = ht.array([2, 3, 4], split=0)
-    >>> ht.vstack((a,b))
-    [0] tensor([[1, 2],
-    [0]         [2, 3]])
-    [1] tensor([[3],
-    [1]         [4]])
+    >>> ht.vstack((a,b))._DNDarray__array
+    [0/1] tensor([[1, 2],
+    [0/1]         [2, 3]])
+    [1/1] tensor([[3],
+    [1/1]         [4]])
     >>> a = ht.array([[1], [2], [3]], split=0)
     >>> b = ht.array([[2], [3], [4]], split=0)
-    >>> ht.vstack((a,b))
-    [0] tensor([[1],
-    [0]         [2],
-    [0]         [3]])
-    [1] tensor([[2],
-    [1]         [3],
-    [1]         [4]])
-
+    >>> ht.vstack((a,b))._DNDarray__array
+    [0/1] tensor([[1],
+    [0/1]         [2],
+    [0/1]         [3]])
+    [1/1] tensor([[2],
+    [1/1]         [3],
+    [1/1]         [4]])
     """
     tup = list(tup)
     for cn, arr in enumerate(tup):
@@ -1872,19 +2235,23 @@ def topk(a, k, dim=None, largest=True, sorted=True, out=None):
         if dim == a.split:
             offset, _, _ = a.comm.chunk(shape, a.split)
             indices = indices.clone()
-            indices += torch.tensor(offset * a.comm.rank, dtype=indices.dtype)
+            indices += torch.tensor(
+                offset * a.comm.rank, dtype=indices.dtype, device=indices.device
+            )
 
         local_shape = list(result.shape)
         local_shape_len = len(shape)
 
-        metadata = torch.tensor([k, dim, largest, sorted, local_shape_len, *local_shape])
+        metadata = torch.tensor(
+            [k, dim, largest, sorted, local_shape_len, *local_shape], device=indices.device
+        )
         send_buffer = torch.cat(
             (metadata.double(), result.double().flatten(), indices.flatten().double())
         )
 
         return send_buffer
 
-    gres = operations.__reduce_op(
+    gres = _operations.__reduce_op(
         a,
         local_topk,
         MPI_TOPK,